requirements.txt
IMAPClient==2.1.0
pyzmail36==1.0.4
six==1.15.0
